// src/app/jobs/companies-network/formData.interface.ts
export class personal{
    firstName: string;
    lastName: string;
    age: string;
    phone: string;
    email: string;
    description: string;
}

export class educational{
    lastDegree: string;
    institute: string;
    major: string;
    yearPassed: string;
}

export class experience{
    category: string;
    durration: string;
    company: string;
    detail: string;
}